move files outside of the /api directory
